// types/sanityTypes.ts
export interface Review {
  title?: string;
  description?: string;
  reviewer?: string;
  image?: SanityImageSource;
}

export interface Reviews {
  title?: string;
  textColor?: HighlightItem[];
  description?: string;
  reviews?: Review[];
}